Fix: userId 타입 변경

+++ Back/petPlace/src/main/java/com/minjeok4go/petplace/feed/entity/Feed.java
@@ -26,7 +26,7 @@ public class Feed {
     private String content;
 
     @Column(name = "user_id", nullable = false)
-    private Integer userId;  // users.id는 INT 타입
+    private Long userId;  // users.id는 INT 타입
 
     @Column(name = "user_nick", nullable = false)
     private String userNick;
